Backend/requirements.txt
Django==3
django-allauth
django-cors-headers
django-filter
django-heroku
django-nested-admin
django-redis
django-rest-auth
djangorestframework
coreapi
drf-yasg
gunicorn
Pillow
psycopg2-binary
python-decouple
redis
Faker
whitenoise
celery
django-celery-results
django-cloudinary-storage
numpy
django-map-widgets
django-leaflet
opencv-python